Refactor `fs` utils
