+++ src/utils/fs/fileUtils.ts
@@ -1,0 +1,112 @@
+import * as fs from "fs";
+import * as path from "path";
+
+import { wrapNonError } from "../errorsUtils";
+import { illegalState } from "../throwErrors";
+
+import { createDirectory } from "./directoryUtils";
+import { isDirectory } from "./fileTypeCheckers";
+import { getDirectoryPath, joinPaths, parsePath } from "./pathUtils";
+
+export const defaultEncoding = "utf-8";
+
+export function exists(inputPath: string): boolean {
+    return fs.existsSync(inputPath);
+}
+
+export function readFile<T>(
+    filePath: string,
+    onError: (error: Error) => T
+): string | T {
+    try {
+        return fs.readFileSync(filePath, defaultEncoding);
+    } catch (e) {
+        return handleThrownObject(e, onError);
+    }
+}
+
+export function writeToFile<T>(
+    text: string,
+    filePath: string,
+    onError: (error: Error) => T,
+    createParentDirectories: boolean = false
+): T | undefined {
+    try {
+        if (createParentDirectories) {
+            createDirectory(false, getDirectoryPath(filePath));
+        }
+        fs.writeFileSync(filePath, text, defaultEncoding);
+        return undefined;
+    } catch (e) {
+        return handleThrownObject(e, onError);
+    }
+}
+
+export function appendToFile<T>(
+    text: string,
+    filePath: string,
+    onError: (e: any) => T
+): T | undefined {
+    try {
+        fs.appendFileSync(filePath, text, defaultEncoding);
+        return undefined;
+    } catch (e) {
+        return handleThrownObject(e, onError);
+    }
+}
+
+function handleThrownObject<T>(e: any, onError: (e: any) => T): T {
+    if (e instanceof Error) {
+        return onError(e);
+    }
+    return onError(
+        wrapNonError(e, "non-`Error` object is thrown inside `writeToFile`")
+    );
+}
+
+export function clearFile(filePath: string) {
+    fs.writeFileSync(filePath, "");
+}
+
+export function deleteFile(filePath: string) {
+    fs.rmSync(filePath, { force: true });
+}
+
+export function copyFile(
+    sourceFilePath: string,
+    destPath: string,
+    throwOnExisting: boolean
+): string {
+    let destFilePath = isDirectory(destPath)
+        ? joinPaths(destPath, parsePath(sourceFilePath).base)
+        : destPath;
+    fs.copyFileSync(
+        sourceFilePath,
+        destFilePath,
+        throwOnExisting ? fs.constants.COPYFILE_EXCL : undefined
+    );
+    return destFilePath;
+}
+
+export type FileCreationModeOnExisting = "throw" | "clear" | "return";
+
+export function createFileWithParentDirectories(
+    mode: FileCreationModeOnExisting,
+    filePath: string
+): string {
+    if (fs.existsSync(filePath)) {
+        switch (mode) {
+            case "throw":
+                illegalState(`failed to create ${filePath}: it already exists`);
+            case "clear":
+                clearFile(filePath);
+                return filePath;
+            case "return":
+                return filePath;
+        }
+    }
+    const parentDirPath = path.dirname(filePath);
+    createDirectory(false, parentDirPath);
+    clearFile(filePath);
+    return filePath;
+}
